backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
openai==1.3.0
sqlmodel==0.0.14
alembic==1.12.1
pydantic==2.5.0
kafka-python==2.0.2
aiokafka==0.8.1
